+++ sdk/apimanagement/api-management-custom-widgets-scaffolder/tsconfig.src.json
@@ -2,5 +2,6 @@
   "extends": "../../../tsconfig.lib.json",
   "compilerOptions": {
     "skipLibCheck": true
-  }
+  },
+  "exclude": ["./src/templates"]
 }
