+++ sdk/apimanagement/api-management-custom-widgets-tools/src/node/deploy.ts
@@ -3,35 +3,11 @@
 
 import type { Config } from "./CustomWidgetBlobService.js";
 import CustomWidgetBlobService from "./CustomWidgetBlobService.js";
-import type { InteractiveBrowserCredentialNodeOptions } from "@azure/identity";
 import { APIM_CONFIG_FILE_NAME } from "../paths.js";
 import fs from "node:fs";
 import getStorageSasUrl from "./getStorageSasUrl.js";
 import readdir from "./readdir.js";
-
-/**
- * resourceId - resource ID of API Management service "subscriptions/[subscription-id]/resourceGroups/[resource-group-name]/providers/Microsoft.ApiManagement/service/[service-name]"
- * managementApiEndpoint - URL with protocol (e.g. https://management.azure.com)
- * apiVersion - optional to override default (e.g. "2019-01-01")
- * tokenOverride - optional, provides token to use for auth, instead of 'az login' approach
- */
-export type ServiceInformation = {
-  resourceId: string;
-  managementApiEndpoint: string;
-  apiVersion?: string;
-  tokenOverride?: string;
-};
-
-/**
- * Optional options object for configuring the deployment function.
- *
- * @param rootLocal - optional, root of the local folder with compiled project to be exported (by default "./dist")
- * @param interactiveBrowserCredentialOptions - options for InteractiveBrowserCredential for Node or InBrowser from \@azure/identity
- */
-export type DeployConfig = {
-  rootLocal?: string;
-  interactiveBrowserCredentialOptions?: InteractiveBrowserCredentialNodeOptions;
-};
+import type { DeployConfig, ServiceInformation } from "./types.js";
 
 /**
  * Deploys everything from /dist folder to the API Management DevPortals' blob storage.
