+++ sdk/apimanagement/api-management-custom-widgets-tools/src/node/getStorageSasUrl.ts
@@ -5,7 +5,7 @@ import {
   InteractiveBrowserCredential,
   type InteractiveBrowserCredentialNodeOptions as IBCNOptions,
 } from "@azure/identity";
-import type { ServiceInformation } from "./deploy.js";
+import type { ServiceInformation } from "./types.js";
 import { getClient } from "@azure-rest/core-client";
 
 async function getAccessToken(
