--- sdk/communication/communication-chat/src/shims.d.ts
@@ -1,11 +1,0 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT License.
-
-// d.ts shims provide types for things we use internally but are not part
-// of this package's surface area.
-
-// Shim for DOM's navigator's product status
-interface Navigator {
-    readonly product: string;
-}
-declare let navigator: Navigator;
