+++ sdk/identity/identity-cache-persistence/tsconfig.json
@@ -3,5 +3,8 @@
     { "path": "./tsconfig.src.json" },
     { "path": "./tsconfig.samples.json" },
     { "path": "./tsconfig.test.json" }
-  ]
+  ],
+  "compilerOptions": {
+    "esModuleInterop": true
+  }
 }
