+++ sdk/monitor/monitor-query/src/internal/modelConverters.ts
@@ -26,20 +26,14 @@ import type {
 
 import type { MetricNamespace as GeneratedMetricNamespace } from "../generated/metricsnamespaces/src/index.js";
 import { formatPreferHeader } from "./util.js";
-
 import type {
   ListMetricDefinitionsOptions,
-  LogsQueryBatchResult,
-  LogsTable,
-  MetricsQueryOptions,
-  MetricsQueryResult,
-  QueryBatch,
-} from "../../src/index.js";
-import type {
   Metric,
   MetricAvailability,
   MetricDefinition,
   MetricNamespace,
+  MetricsQueryOptions,
+  MetricsQueryResult,
   TimeSeriesElement,
 } from "../models/publicMetricsModels.js";
 import { createMetricsQueryResult, getMetricByName } from "../models/publicMetricsModels.js";
@@ -50,9 +44,12 @@ import {
 } from "../timespanConversion.js";
 import type {
   LogsErrorInfo,
+  LogsQueryBatchResult,
   LogsQueryError,
   LogsQueryPartialResult,
   LogsQuerySuccessfulResult,
+  LogsTable,
+  QueryBatch,
 } from "../models/publicLogsModels.js";
 import { LogsQueryResultStatus } from "../models/publicLogsModels.js";
 import type {
