+++ sdk/openai/openai/tsconfig.json
@@ -3,5 +3,9 @@
     { "path": "./tsconfig.src.json" },
     { "path": "./tsconfig.samples.json" },
     { "path": "./tsconfig.test.json" }
-  ]
+  ],
+  "compilerOptions": {
+    "target": "ES2017",
+    "moduleResolution": "nodenext"
+  }
 }
