+++ sdk/schemaregistry/schema-registry-avro/src/index.ts
@@ -1,6 +1,11 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
+declare global {
+    interface Blob {}
+  }
+  
+
 export { AvroSerializer } from "./avroSerializer.js";
 
 export * from "./models.js";
