--- sdk/schemaregistry/schema-registry-avro/types/schema-registry-avro.shims.d.ts
@@ -1,8 +1,0 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT license.
-
-declare global {
-  interface Blob {}
-}
-
-export * from "./schema-registry-avro";
