+++ sdk/servicebus/service-bus/src/util/runtimeInfo-browser.mts
@@ -1,18 +1,6 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-declare global {
-  interface Navigator {
-    platform: string;
-  }
-}
-
-declare const self: {
-  navigator: {
-    platform: string;
-  };
-};
-
 /**
  * @internal
  */
