+++ sdk/servicebus/service-bus/tsconfig.src.json
@@ -1,6 +1,7 @@
 {
   "extends": "../../../tsconfig.lib.json",
   "compilerOptions": {
-    "skipLibCheck": true
+    "skipLibCheck": true,
+    "lib": ["DOM"]
   }
 }
